+++ mRemoteV1/CS/Connection/Connection.Protocol.HTTPS.cs
@@ -1,4 +1,3 @@
-// VBConversions Note: VB project level imports
 using System.Collections.Generic;
 using System;
 using AxWFICALib;
@@ -9,29 +8,27 @@ using AxMSTSCLib;
 using Microsoft.VisualBasic;
 using System.Collections;
 using System.Windows.Forms;
-// End of VB project level imports
 
 
 namespace mRemoteNG.Connection.Protocol
 {
 	public class HTTPS : HTTPBase
-			{
+	{
 				
-				public HTTPS(RenderingEngine RenderingEngine) : base(RenderingEngine)
-				{
-				}
+		public HTTPS(RenderingEngine RenderingEngine) : base(RenderingEngine)
+		{
+		}
 				
-				public override void NewExtended()
-				{
-					base.NewExtended();
-					
-					httpOrS = "https";
-					defaultPort = Defaults.Port;
-				}
+		public override void NewExtended()
+		{
+			base.NewExtended();
+			httpOrS = "https";
+			defaultPort = (int)Defaults.Port;
+		}
 				
-				public enum Defaults
-				{
-					Port = 443
-				}
-			}
+		public enum Defaults
+		{
+			Port = 443
+		}
+	}
 }
